separate social schema and social types regex
